requirements.txt
attrs==19.3.0
pygatt=4.0.5
pexpect=4.8.0